+++ src/plugins/local-completion.bash
@@ -1,0 +1,10 @@
+#! /bin/bash
+# local bash completion files
+if [[ -d ${HOME}/.bash_completion.d ]]; then
+    for i in ${HOME}/.bash_completion.d/*.*sh; do
+        if [ -r $i ]; then
+            . $i
+        fi
+    done
+    unset i
+fi