+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased][unreleased]
 
+#### Added
+- local-completion plugin
+
 ### 0.1.0 - 2015-10-01
 #### Added
 - package files